Fix selection sort to use float type for temporary variable

+++ 1st_Year/2SEM/DS/a1q3_selection_sort.c
@@ -26,7 +26,7 @@ void selectionSort(float *arr, int arrSize)
 				min = j;
 			}
 		}
-		int temp = arr[min];
+		float temp = arr[min];
 		arr[min] = arr[i];
 		arr[i] = temp;
 	}
